Refactor to update to Jaeger extension project
