--- integration-tests/src/test/resources/testng.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-  ~ Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-  ~
-  ~ WSO2 Inc. licenses this file to you under the Apache License,
-  ~ Version 2.0 (the "License"); you may not use this file except
-  ~ in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~    http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing,
-  ~ software distributed under the License is distributed on an
-  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  ~ KIND, either express or implied.  See the License for the
-  ~ specific language governing permissions and limitations
-  ~ under the License.
-  -->
-
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
-<suite name="Ballerina Observability Extension Tests Suite" parallel="false">
-    <listeners>
-        <listener class-name="org.ballerinalang.test.listener.TestExecutionListener"/>
-        <listener class-name="org.ballerinalang.test.listener.TestNGListener"/>
-    </listeners>
-
-    <test name="ballerina-observability-tests" parallel="false">
-        <classes>
-            <class name="org.ballerinalang.observe.PrometheusMetricsTestCase"/>
-            <class name="org.ballerinalang.observe.JaegerTracesTestCase"/>
-        </classes>
-    </test>
-</suite>
