+++ jaeger-extension-ballerina/Ballerina.toml
@@ -1,0 +1,56 @@
+# Copyright (c) 2020 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+#
+# WSO2 Inc. licenses this file to you under the Apache License,
+# Version 2.0 (the "License"); you may not use this file except
+# in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
+[package]
+org = "ballerinax"
+name = "jaeger"
+version = "@toml.version@"
+
+[[platform.java11.dependency]]
+path = "../jaeger-extension-native/build/libs/jaeger-extension-native-@project.version@.jar"
+groupId = "ballerina"
+artifactId = "jaeger-extension-native"
+version = "@project.version@"
+
+[[platform.java11.dependency]]
+path = "./lib/jaeger-core-@jaeger.version@.jar"
+groupId = "io.jaegertracing"
+artifactId = "jaeger-core"
+version = "@jaeger.version@"
+
+[[platform.java11.dependency]]
+path = "./lib/jaeger-thrift-@jaeger.version@.jar"
+groupId = "io.jaegertracing"
+artifactId = "jaeger-thrift"
+version = "@jaeger.version@"
+
+[[platform.java11.dependency]]
+path = "./lib/libthrift-@libthrift.version@.jar"
+groupId = "org.apache.thrift"
+artifactId = "libthrift"
+version = "@libthrift.version@"
+
+[[platform.java11.dependency]]
+path = "./lib/okhttp-@okhttp.version@.jar"
+groupId = "com.squareup.okhttp3"
+artifactId = "okhttp"
+version = "@okhttp.version@"
+
+[[platform.java11.dependency]]
+path = "./lib/okio-@okio.version@.jar"
+groupId = "com.squareup.okio"
+artifactId = "okio"
+version = "@okio.version@"
