+++ jaeger-extension-native/src/main/java/module-info.java
@@ -1,0 +1,8 @@
+module io.ballerina.observe.trace.extension.jaeger {
+    requires io.ballerina.runtime;
+    requires io.ballerina.config;
+    requires opentracing.api;
+    requires jaeger.core;
+
+    exports io.ballerina.observe.trace.jaeger;
+}
