+++ jaeger-extension-native/src/main/resources/jaeger-tracer.properties
