+++ jaeger-extension-tests/src/test/resources/logging.properties
