--- prometheus-extension/ballerina/Ballerina.toml
@@ -1,46 +1,0 @@
-# Copyright (c) 2020 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-#
-# WSO2 Inc. licenses this file to you under the Apache License,
-# Version 2.0 (the "License"); you may not use this file except
-# in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-[package]
-org = "ballerina"
-name = "prometheus"
-version = "@toml.version@"
-
-[[dependency]]
-org = "ballerina"
-name = "config"
-version = "@stdlib.config.version@"
-
-[[dependency]]
-org = "ballerina"
-name = "stringutils"
-version = "@stdlib.stringutils.version@"
-
-[[dependency]]
-org = "ballerina"
-name = "http"
-version = "@stdlib.http.version@"
-
-[[dependency]]
-org = "ballerina"
-name = "observe"
-version = "@stdlib.observe.version@"
-
-[[platform.java11.dependency]]
-path = "../native/build/libs/prometheus-extension-native-@project.version@.jar"
-groupId = "ballerina"
-artifactId = "prometheus-extension-native"
-version = "@project.version@"
