--- prometheus-extension/native/src/main/java/module-info.java
@@ -1,6 +1,0 @@
-module io.ballerina.observe.metrics.prometheus {
-    requires io.ballerina.runtime;
-    requires io.ballerina.config;
-
-    exports org.ballerinalang.observe.metrics.prometheus;
-}
